prelim attempt to make ALGO work with STRAT
This code has a problem: when ALGO returns to the
STRAT to issue the orders, any identifiers in
the order cause a segfault.  It appears that the
values are correct in the symbol table, but not
in the ast_order_item.  Don't know why it's not
getting updated.  Otherwise...not so bad.